requirements.txt
fastapi==0.114.0
uvicorn==0.30.1
google-auth==2.29.0
google-api-python-client==2.100.0
jinja2==3.1.3
python-multipart==0.0.9
sqlalchemy==2.0.30